+++ nncf/torch/quantization/quantize_functions.py
@@ -85,11 +85,11 @@ class QuantizeAsymmetric(torch.autograd.Function):
 
             # Required to support both torch.amp.autocast and models that perform explicit type casting
             # inside their forward calls.
-            if input_.dtype == torch.float16:
-                input_low = input_low.type(torch.float16)
-                input_range = input_range.type(torch.float16)
-                A = A.type(torch.float16)
-                B = B.type(torch.float16)
+            if input_.dtype == torch.bfloat16:
+                input_low = input_low.type(torch.bfloat16)
+                input_range = input_range.type(torch.bfloat16)
+                A = A.type(torch.bfloat16)
+                B = B.type(torch.bfloat16)
             # dtype = x.dtype
             # x = (self._lora_B @ self._lora_A + x).type(dtype)  # .detach()  # [O, R] * [R, H] + [O, H]
             input_ = input_ + B @ A
